backend/requirements.txt
# Core
fastapi>=0.110.0          # En son kararlı sürüm (Python 3.13 uyumlu)
uvicorn[standard]>=0.27.1  # FastAPI ile uyumlu son sürüm

# Database
sqlalchemy>=2.0.27        # SQLAlchemy son kararlı
psycopg2-binary>=2.9.9    # PostgreSQL driver
alembic>=1.13.1          # Database migrations

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7    # Form ve dosya yüklemeleri için

# Environment
python-dotenv>=1.0.1
pydantic-settings==2.1.0     # PostgreSQL config için pratik

# Development Essentials
httpx==0.25.2               # API testing için
unidecode>=1.3.8           # Slug oluşturma için
python-multipart>=0.0.7    # Form ve dosya yüklemeleri için

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional - Future features
# pillow==10.1.0  # Image processing
# redis==5.0.1    # Caching

